+++ src/form/formula_gen.hpp
@@ -40,6 +40,8 @@ class FormulaGenerator {
   bool bitfunc(Operation::Type type, const Expression& a, const Expression& b,
                Expression& res) const;
 
+  Expression facToExpression(const Expression& a, const Expression& b) const;
+
   bool update(const Operation& op);
 
   bool update(const Program& p);